+++ src/object.cpp
@@ -1144,6 +1144,41 @@ int Object::FindAllBetween(FunctorParams *functorParams)
     return FUNCTOR_CONTINUE;
 }
 
+int Object::FindAllReferencedObjects(FunctorParams *functorParams)
+{
+    FindAllReferencedObjectsParams *params = dynamic_cast<FindAllReferencedObjectsParams *>(functorParams);
+    assert(params);
+
+    if (this->HasInterface(INTERFACE_LINKING)) {
+        LinkingInterface *interface = this->GetLinkingInterface();
+        assert(interface);
+        if (interface->GetNextLink()) params->m_elements->push_back(interface->GetNextLink());
+        if (interface->GetSameasLink()) params->m_elements->push_back(interface->GetSameasLink());
+    }
+    if (this->HasInterface(INTERFACE_PLIST)) {
+        PlistInterface *interface = this->GetPlistInterface();
+        assert(interface);
+        for (auto &object : *interface->GetRefs()) {
+            params->m_elements->push_back(object);
+        }
+    }
+    if (this->HasInterface(INTERFACE_TIME_POINT)) {
+        TimePointInterface *interface = this->GetTimePointInterface();
+        assert(interface);
+        if (interface->GetStart() && !interface->GetStart()->Is(TIMESTAMP_ATTR))
+            params->m_elements->push_back(interface->GetStart());
+    }
+    if (this->HasInterface(INTERFACE_TIME_SPANNING)) {
+        TimeSpanningInterface *interface = this->GetTimeSpanningInterface();
+        assert(interface);
+        if (interface->GetEnd() && !interface->GetEnd()->Is(TIMESTAMP_ATTR))
+            params->m_elements->push_back(interface->GetEnd());
+    }
+
+    // continue until the end
+    return FUNCTOR_CONTINUE;
+}
+
 int Object::ConvertToCastOffMensural(FunctorParams *functorParams)
 {
     ConvertToCastOffMensuralParams *params = dynamic_cast<ConvertToCastOffMensuralParams *>(functorParams);