+++ tools/main.cpp
@@ -116,6 +116,9 @@ void display_usage()
     std::cout << " -t, --to <s>          Select output format: mei, pb-mei, svg, or midi (default is svg)" << std::endl;
     std::cout << " -v, --version         Display the version number" << std::endl;
     std::cout << " -x, --xml-id-seed <i> Seed the random number generator for XML IDs" << std::endl;
+    
+    std::cout << std::endl << "Additional long options" << std::endl;
+    std::cout << "--remove-ids           Remove in the MEI output XML IDs that are not referenced " << std::endl;
 
     vrv::Options options;
     std::vector<vrv::OptionGrp *> *grp = options.GetGrps();
@@ -193,6 +196,7 @@ int main(int argc, char **argv)
     std::string outfile;
     std::string outformat = "svg";
     bool std_output = false;
+    bool remove_ids = false;
 
     int all_pages = 0;
     int page = 1;
@@ -221,14 +225,8 @@ int main(int argc, char **argv)
             { "to", required_argument, 0, 't' },
             { "version", no_argument, 0, 'v' },
             { "xml-id-seed", required_argument, 0, 'x' },
-            // deprecated - some use undocumented short options to catch them as such
-            { "border", required_argument, 0, 'b' },
-            { "format", required_argument, 0, 'd' },
-            { "ignore-layout", no_argument, 0, 'd' },
-            { "no-footer", no_argument, 0, 'd' },
-            { "no-header", no_argument, 0, 'd' },
-            { "no-layout", no_argument, 0, 'd' },
-            { "type", required_argument, 0, 'd' },
+            // mei output - long options only
+            { "remove-ids", no_argument, 0, 'm' },
             { 0, 0, 0, 0 }
         };
 
@@ -274,7 +272,7 @@ int main(int argc, char **argv)
     int option_index = 0;
     vrv::Option *opt = NULL;
     vrv::OptionBool *optBool = NULL;
-    while ((c = getopt_long(argc, argv, "ab:f:hino:p:r:s:t:vx:", long_options, &option_index)) != -1) {
+    while ((c = getopt_long(argc, argv, "ab:f:hmo:p:r:s:t:vx:", long_options, &option_index)) != -1) {
         switch (c) {
             case 0:
                 key = long_options[option_index].name;
@@ -307,40 +305,16 @@ int main(int argc, char **argv)
                 options->m_pageMarginTop.SetValue(optarg);
                 break;
 
-            case 'd':
-                if (!strcmp(long_options[option_index].name, "format")) {
-                    vrv::LogWarning("Option --format is deprecated; use --from");
-                    if (!toolkit.SetInputFrom(std::string(optarg))) {
-                        exit(1);
-                    };
-                }
-                else if (!strcmp(long_options[option_index].name, "ignore-layout")) {
-                    vrv::LogWarning("Option --ignore-layout is deprecated; use --breaks auto");
-                    options->m_breaks.SetValue(vrv::BREAKS_auto);
-                }
-                else if (!strcmp(long_options[option_index].name, "no-footer")) {
-                    vrv::LogWarning("Option --no-footer is deprecated; use --footer none");
-                    options->m_footer.SetValue(vrv::FOOTER_none);
-                }
-                else if (!strcmp(long_options[option_index].name, "no-header")) {
-                    vrv::LogWarning("Option --no-header is deprecated; use --header none");
-                    options->m_header.SetValue(vrv::HEADER_none);
-                }
-                else if (!strcmp(long_options[option_index].name, "no-layout")) {
-                    vrv::LogWarning("Option --no-layout is deprecated; use --breaks none");
-                    options->m_breaks.SetValue(vrv::BREAKS_none);
-                }
-                else if (!strcmp(long_options[option_index].name, "type")) {
-                    vrv::LogWarning("Option --type is deprecated; use --to");
-                    outformat = std::string(optarg);
-                    toolkit.SetOutputTo(std::string(optarg));
-                }
-                break;
-
             case 'f':
                 if (!toolkit.SetInputFrom(std::string(optarg))) {
                     exit(1);
                 };
+                break;
+                
+            case 'm':
+                if (!strcmp(long_options[option_index].name, "remove-ids")) {
+                    remove_ids = true;
+                }
                 break;
 
             case 'o': outfile = std::string(optarg); break;
@@ -553,9 +527,10 @@ int main(int argc, char **argv)
     }
     else {
         const char *scoreBased = (outformat == "mei") ? "true" : "false";
+        const char *removeIds = (remove_ids) ? "true" : "false";
         outfile += ".mei";
         if (all_pages) {
-            std::string params = vrv::StringFormat("{'scoreBased': %s}", scoreBased);
+            std::string params = vrv::StringFormat("{'scoreBased': %s, 'removeIds': %s}", scoreBased, removeIds);
             if (std_output) {
                 std::string output;
                 std::cout << toolkit.GetMEI(params);
@@ -569,7 +544,7 @@ int main(int argc, char **argv)
             }
         }
         else {
-            std::string params = vrv::StringFormat("{'scoreBased': %s, 'pageNo': %d}", scoreBased, page);
+            std::string params = vrv::StringFormat("{'scoreBased': %s, 'pageNo': %d, 'removeIds': %s}", scoreBased, page, removeIds);
             if (std_output) {
                 std::cout << toolkit.GetMEI(params);
             }